+++ src/main/java/org/example/bankingsystem/dto/response/ViewAccountResponse.java
@@ -5,14 +5,14 @@ import lombok.Data;
 import org.example.bankingsystem.enums.AccountStatus;
 import org.example.bankingsystem.enums.AccountType;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
 public class ViewAccountResponse {
     private String accountNumber;
-    private Double balance;
+    private BigDecimal balance;
     private AccountType accountType;
     private AccountStatus accountStatus;
     private LocalDateTime dateCreated;
